moving more tests to runtime
