+++ Assets/Mirror/Tests/Common/ClientSceneTests_RegisterPrefabBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8eb53689226946640bc49bb962b13638
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
