+++ Assets/Mirror/Tests/Editor/ClientSceneTests_RegisterPrefab.cs
@@ -95,63 +95,6 @@ namespace Mirror.Tests.ClientSceneTests
         }
 
         [Test]
-        [TestCase(RegisterPrefabOverload.Prefab)]
-        [TestCase(RegisterPrefabOverload.Prefab_SpawnDelegate)]
-        [TestCase(RegisterPrefabOverload.Prefab_SpawnHandlerDelegate)]
-        public void ErrorForEmptyGuid(RegisterPrefabOverload overload)
-        {
-            // setup
-            CreateSceneObject(out GameObject runtimeObject, out NetworkIdentity networkIdentity);
-
-            //test
-            string msg = OverloadWithHandler(overload)
-               ? $"Can not Register handler for '{runtimeObject.name}' because it had empty assetid. If this is a scene Object use RegisterSpawnHandler instead"
-               : $"Can not Register '{runtimeObject.name}' because it had empty assetid. If this is a scene Object use RegisterSpawnHandler instead";
-
-            LogAssert.Expect(LogType.Error, msg);
-            CallRegisterPrefab(runtimeObject, overload);
-
-            // teardown
-            GameObject.DestroyImmediate(runtimeObject);
-        }
-
-        [Test]
-        [TestCase(RegisterPrefabOverload.Prefab_NewAssetId)]
-        public void PrefabNewGuid_AddsRuntimeObjectToDictionary(RegisterPrefabOverload overload)
-        {
-            // setup
-            CreateSceneObject(out GameObject runtimeObject, out NetworkIdentity networkIdentity);
-
-            //test
-            CallRegisterPrefab(runtimeObject, overload);
-
-            Assert.IsTrue(prefabs.ContainsKey(anotherGuid));
-            Assert.AreEqual(prefabs[anotherGuid], runtimeObject);
-
-            Assert.AreEqual(networkIdentity.assetId, anotherGuid);
-
-            // teardown
-            GameObject.DestroyImmediate(runtimeObject);
-        }
-
-        [Test]
-        [TestCase(RegisterPrefabOverload.Prefab_SpawnDelegate_NewAssetId)]
-        [TestCase(RegisterPrefabOverload.Prefab_SpawnHandlerDelegate_NewAssetId)]
-        public void Handler_AddsUnSpawnHandlerToDictionaryForRuntimeObject(RegisterPrefabOverload overload)
-        {
-            // setup
-            CreateSceneObject(out GameObject runtimeObject, out NetworkIdentity networkIdentity);
-
-            //test
-            CallRegisterPrefab(runtimeObject, overload);
-
-            Assert.IsTrue(unspawnHandlers.ContainsKey(anotherGuid));
-
-            // teardown
-            GameObject.DestroyImmediate(runtimeObject);
-        }
-
-        [Test]
         [TestCase(RegisterPrefabOverload.Prefab_NewAssetId)]
         [TestCase(RegisterPrefabOverload.Prefab_SpawnDelegate_NewAssetId)]
         [TestCase(RegisterPrefabOverload.Prefab_SpawnHandlerDelegate_NewAssetId)]
